split src into phase and cphase for dll imports
