+++ src/cphase/stereomatcher/stereomatcher.cs
@@ -1,0 +1,51 @@
+/*!
+ * @authors Ben Knight (bknight@i3drobotics.com)
+ * @date 2021-05-26
+ * @copyright Copyright (c) I3D Robotics Ltd, 2021
+ * 
+ * @file stereomatcher.cs
+ * @brief Stereo Matcher class
+ * @details Create stereo matcher of any type using 
+ * StereoMatcherType. This allows for the same interface to be
+ * used for any stereo matcher. 
+ */
+
+using System;
+using System.Runtime.InteropServices;
+
+namespace I3DR.CPhase
+{
+    //!  Stereo Matcher class
+    /*!
+    Create stereo matcher of any type using 
+    StereoMatcherType. This allows for the same interface to be
+    used for any stereo matcher. 
+    */
+    public class StereoMatcher
+    {
+        //! Imported from Phase C API
+        [DllImport("phase", EntryPoint = "I3DR_CCreateStereoMatcher", CallingConvention = CallingConvention.Cdecl)]
+        private static extern IntPtr CCreateStereoMatcher(StereoMatcherType matcher_type);
+
+        /*!
+        * Create stereo matching from stereo matcher type
+        *
+        * @param matcher_type stereo matcher type to create
+        * @returns stereo matcher object of type specified
+        */
+        public static AbstractStereoMatcher createStereoMatcher(StereoMatcherType matcher_type){
+            if (matcher_type == StereoMatcherType.STEREO_MATCHER_I3DRSGM){
+                return new StereoI3DRSGM(CCreateStereoMatcher(matcher_type));
+            } else if (matcher_type == StereoMatcherType.STEREO_MATCHER_BM){
+                return new StereoBM(CCreateStereoMatcher(matcher_type));
+            } else if (matcher_type == StereoMatcherType.STEREO_MATCHER_SGBM){
+                return new StereoSGBM(CCreateStereoMatcher(matcher_type));
+            } else if (matcher_type == StereoMatcherType.STEREO_MATCHER_HOBM){
+                return new StereoHOBM(CCreateStereoMatcher(matcher_type));
+            } else {
+                throw new ArgumentException(
+                    String.Format("Unsupported matcher type: {0}", matcher_type),"matcher_type");
+            }
+        }
+    }
+}
