+++ src/cphase/types/stereo.cs
@@ -1,0 +1,17 @@
+/*!
+ * @authors Ben Knight (bknight@i3drobotics.com)
+ * @date 2021-05-26
+ * @copyright Copyright (c) I3D Robotics Ltd, 2021
+ * 
+ * @file common.cs
+ * @brief Common types.
+ * @brief Commmon types used in Phase
+ */
+
+using System;
+using System.Runtime.InteropServices;
+using System.Runtime.ExceptionServices;
+
+namespace I3DR.CPhase
+{
+}
